+++ nuget-client/BrevgeneratorClientAdhocTest/Program.cs
@@ -22,13 +22,13 @@ static class Program
         // ---------- Klient 1 via DI ----------
         var serviceCollection = new ServiceCollection();
         serviceCollection.AddSingleton(new BrevgeneratorConfig(apiUrl));
-        serviceCollection.AddSingleton<IBrevgeneratorKlient>(sp => new BrevgeneratorKlient(
+        serviceCollection.AddSingleton<IBrevgeneratorClient>(sp => new BrevgeneratorClient(
             sp.GetRequiredService<BrevgeneratorConfig>(),
-            BrevgeneratorKlient.AuthMode.ApiKey,
+            BrevgeneratorClient.AuthMode.ApiKey,
             apiKeyFactory: async () => "foo"
         ));
         var serviceProvider = serviceCollection.BuildServiceProvider();
-        var client1 = serviceProvider.GetRequiredService<IBrevgeneratorKlient>();
+        var client1 = serviceProvider.GetRequiredService<IBrevgeneratorClient>();
 
         // ---------- Klient 2 direkte konstruksjon ----------
         // var client2 = new BrevgeneratorKlient(